+++ src/mod/Stats/PlayerStats.h
@@ -14,7 +14,7 @@ private:
     mce::UUID                  mUuid;
     std::string                mXuid;
     std::string                mName;
-    int                        mSneakingStartTick;
+    uint64_t                   mSneakingStartTick;
     std::filesystem::path      getPath();
     // void                  parseData(std::string const& data);
 
@@ -25,9 +25,9 @@ public:
     nlohmann::json getJson();
     // nlohmann::json getJsonStatsData(StatsDataType type);
     bool saveData();
-    void addStats(StatsDataType type, std::string key, int value = 1);
-    void addCustomStats(CustomType type, int value = 1);
-    void resetCustomStats(CustomType type, int value = 0);
+    void addStats(StatsDataType type, std::string key, uint64_t value = 1);
+    void addCustomStats(CustomType type, uint64_t value = 1);
+    void resetCustomStats(CustomType type, uint64_t value = 0);
     void startSneaking();
     void addSneakTick();
 };
